packages/garagegate/pyproject.toml
[tool.poetry]
name = "garagegate"
version = "0.1.4"
description = ""
authors = ["Mariusz Kabała <mariusz@kabala.waw.pl>"]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.5.0"
requests = "*"
python-json-logger = "^2.0.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
start = 'garagegate:start'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[package]
include = ["*.py"]
